review comments and minor refactor
