+++ yb-voyager/cmd/importSchemaYugabyteDB.go
@@ -20,7 +20,10 @@ import (
 	"fmt"
 	"path/filepath"
 	"strings"
+	"time"
 
+	"github.com/fatih/color"
+	"github.com/jackc/pgx/v4"
 	log "github.com/sirupsen/logrus"
 	"golang.org/x/exp/slices"
 
@@ -28,7 +31,7 @@ import (
 )
 
 var deferredSqlStmts []sqlInfo
-var failedSqlStmts []string
+var finalFailedSqlStmts []string
 
 func importSchemaInternal(exportDir string, importObjectList []string,
 	skipFn func(string, string) bool) {
@@ -41,6 +44,124 @@ func importSchemaInternal(exportDir string, importObjectList []string,
 		executeSqlFile(importObjectFilePath, importObjectType, skipFn)
 	}
 
+}
+
+func executeSqlFile(file string, objType string, skipFn func(string, string) bool) {
+	log.Infof("Execute SQL file %q on target %q", file, tconf.Host)
+	conn := newTargetConn()
+
+	defer func() {
+		if conn != nil {
+			conn.Close(context.Background())
+		}
+	}()
+
+	sqlInfoArr := parseSqlFileForObjectType(file, objType)
+	for _, sqlInfo := range sqlInfoArr {
+		if conn == nil {
+			conn = newTargetConn()
+		}
+
+		setOrSelectStmt := strings.HasPrefix(strings.ToUpper(sqlInfo.stmt), "SET ") ||
+			strings.HasPrefix(strings.ToUpper(sqlInfo.stmt), "SELECT ")
+		if !setOrSelectStmt && skipFn != nil && skipFn(objType, sqlInfo.stmt) {
+			continue
+		}
+
+		if objType == "TABLE" {
+			stmt := strings.ToUpper(sqlInfo.stmt)
+			skip := strings.Contains(stmt, "ALTER TABLE") && strings.Contains(stmt, "REPLICA IDENTITY")
+			if skip {
+				//skipping DDLS like ALTER TABLE ... REPLICA IDENTITY .. as this is not supported in YB
+				log.Infof("Skipping DDL: %s", sqlInfo.stmt)
+				continue
+			}
+		}
+
+		err := executeSqlStmtWithRetries(&conn, sqlInfo, objType)
+		if err != nil {
+			conn.Close(context.Background())
+			conn = nil
+		}
+	}
+}
+
+func executeSqlStmtWithRetries(conn **pgx.Conn, sqlInfo sqlInfo, objType string) error {
+	var err error
+	log.Infof("On %s run query:\n%s\n", tconf.Host, sqlInfo.formattedStmt)
+	for retryCount := 0; retryCount <= DDL_MAX_RETRY_COUNT; retryCount++ {
+		if retryCount > 0 { // Not the first iteration.
+			log.Infof("Sleep for 5 seconds before retrying for %dth time", retryCount)
+			time.Sleep(time.Second * 5)
+			log.Infof("RETRYING DDL: %q", sqlInfo.stmt)
+		}
+
+		if bool(flagPostSnapshotImport) && strings.Contains(objType, "INDEX") {
+			err = beforeIndexCreation(sqlInfo, conn, objType)
+			if err != nil {
+				return fmt.Errorf("before index creation: %w", err)
+			}
+		}
+		_, err = (*conn).Exec(context.Background(), sqlInfo.formattedStmt)
+		if err == nil {
+			utils.PrintSqlStmtIfDDL(sqlInfo.stmt, utils.GetObjectFileName(filepath.Join(exportDir, "schema"), objType))
+			return nil
+		}
+
+		log.Errorf("DDL Execution Failed for %q: %s", sqlInfo.formattedStmt, err)
+		if strings.Contains(strings.ToLower(err.Error()), "conflicts with higher priority transaction") {
+			// creating fresh connection
+			(*conn).Close(context.Background())
+			*conn = newTargetConn()
+			continue
+		} else if strings.Contains(strings.ToLower(err.Error()), strings.ToLower(SCHEMA_VERSION_MISMATCH_ERR)) &&
+			(objType == "INDEX" || objType == "PARTITION_INDEX") { // retriable error
+			// creating fresh connection
+			(*conn).Close(context.Background())
+			*conn = newTargetConn()
+
+			// Extract the schema name and add to the index name
+			fullyQualifiedObjName, err := getIndexName(sqlInfo.stmt, sqlInfo.objName)
+			if err != nil {
+				utils.ErrExit("extract qualified index name from DDL [%v]: %v", sqlInfo.stmt, err)
+			}
+
+			// DROP INDEX in case INVALID index got created
+			// `err` is already being used for retries, so using `err2`
+			err2 := dropIdx(*conn, fullyQualifiedObjName)
+			if err2 != nil {
+				return fmt.Errorf("drop invalid index %q: %w", fullyQualifiedObjName, err2)
+			}
+			continue
+		} else if missingRequiredSchemaObject(err) {
+			log.Infof("deffering execution of SQL: %s", sqlInfo.formattedStmt)
+			deferredSqlStmts = append(deferredSqlStmts, sqlInfo)
+		} else if isAlreadyExists(err.Error()) {
+			// pg_dump generates `CREATE SCHEMA public;` in the schemas.sql. Because the `public`
+			// schema already exists on the target YB db, the create schema statement fails with
+			// "already exists" error. Ignore the error.
+			if bool(tconf.IgnoreIfExists) || strings.EqualFold(strings.Trim(sqlInfo.stmt, " \n"), "CREATE SCHEMA public;") {
+				err = nil
+			}
+		}
+		break // no more iteration in case of non retriable error
+	}
+	if err != nil {
+		if missingRequiredSchemaObject(err) {
+			// Do nothing
+		} else {
+			utils.PrintSqlStmtIfDDL(sqlInfo.stmt, utils.GetObjectFileName(filepath.Join(exportDir, "schema"), objType))
+			color.Red(fmt.Sprintf("%s\n", err.Error()))
+			if tconf.ContinueOnError {
+				log.Infof("appending stmt to failedSqlStmts list: %s\n", utils.GetSqlStmtToPrint(sqlInfo.stmt))
+				errString := "/*\n" + err.Error() + "\n*/\n"
+				finalFailedSqlStmts = append(finalFailedSqlStmts, errString+sqlInfo.formattedStmt)
+			} else {
+				utils.ErrExit("error: %s\n", err)
+			}
+		}
+	}
+	return err
 }
 
 /*
@@ -95,7 +216,7 @@ func importDeferredStatements() {
 			break
 		}
 	}
-	failedSqlStmts = append(failedSqlStmts, finalFailedDeferredStmts...)
+	finalFailedSqlStmts = append(finalFailedSqlStmts, finalFailedDeferredStmts...)
 }
 
 func applySchemaObjectFilterFlags(importObjectOrderList []string) []string {
